user-premium

+++ src/public/js/main.js
@@ -1,5 +1,7 @@
 const socket = io();
 
+socket.emit("getProducts")
+
 socket.on("products", (data) => {
     renderProductos(data);
 });
@@ -8,30 +10,37 @@ const renderProductos = (products) => {
     const productsContainer = document.getElementById("productsContainer");
     productsContainer.innerHTML = "";
 
-    products.forEach(item => {
-        const card = document.createElement("div");
-        card.classList = "card";
-        card.style = "width: 18rem";
-        card.innerHTML = `  <div class="card-body">
-                                <p class="card-title"> ID: ${item._id}</p>
-                                <h2 class="card-title"> Titulo: ${item.title}</h2>
-                                <p class="card-text"> Precio: ${item.price} </p>
-                                <button class="btn btn-primary">Eliminar</button>
-                            </div>
-                        `
-        productsContainer.appendChild(card);
+    if(products.length !== 0){
 
-        card.querySelector("button").addEventListener("click", () => {
-            deleteProduct(item._id);
+        products.forEach(item => {
+            const card = document.createElement("div");
+            card.classList = "card";
+            card.style = "width: 18rem";
+            card.innerHTML = `  <div class="card-body">
+                                    <p class="card-title"> ID: ${item._id}</p>
+                                    <h2 class="card-title"> Titulo: ${item.title}</h2>
+                                    <p class="card-text"> Precio: ${item.price} </p>
+                                    <button class="btn btn-primary">Eliminar</button>
+                                </div>
+                            `
+            productsContainer.appendChild(card);
+
+            card.querySelector("button").addEventListener("click", () => {
+                deleteProduct(item._id);
+            });
         });
-    });
+    }else {
+        const messageWithoutProducts = document.createElement("p");
+        messageWithoutProducts.textContent = "Todavía no has agregado ningún producto"
+        productsContainer.appendChild(messageWithoutProducts)
+    }
 };
 
 const deleteProduct = (id) => {
     socket.emit("deleteProduct", id);
 };
 
-const sendButton = document.getElementById("btnSend").addEventListener("click", () => addProduct())
+
 
 const addProduct = () => {
     const product = {
@@ -46,3 +55,8 @@ const addProduct = () => {
     };
     socket.emit("addProduct", product);
 };
+
+const sendButton = document.getElementById("btnSend").addEventListener("click", (event) => {
+    event.preventDefault();
+    addProduct()
+})
